requirements.txt
streamlit
pypdf
python-docx
numpy
scikit-learn
google-generativeai
python-dotenv